More logs

+++ src/app/(training)/(email-callback)/delete-account/actions.ts
@@ -5,10 +5,11 @@ import { redirect } from "next/navigation";
 
 import type { MessageDescriptor } from "@lingui/core";
 import { msg } from "@lingui/core/macro";
+import { logger } from "better-auth";
 
 import { auth } from "~/lib/auth";
 import { getAuthError } from "~/lib/auth/errors";
-import { verifyPassword } from "~/lib/user";
+import { getSessionUser, verifyPassword } from "~/lib/user";
 
 export async function deleteAccount(
   password: string,
@@ -16,8 +17,13 @@ export async function deleteAccount(
 ): Promise<MessageDescriptor | undefined> {
   if (!token) return msg`Token non valido`;
 
+  const user = await getSessionUser();
+  if (!user) return msg`Utente non trovato`;
+
   const err = await verifyPassword(password);
   if (err) return err;
+
+  logger.warn(`Deleting user ${user.username}`);
 
   try {
     await auth.api.deleteUserCallback({
